1. Filter possible before loading data 2. Search only the current data

+++ Timeline.py
@@ -57,6 +57,7 @@ class Main(tk.Frame):
         self.inner_fields_frame_1 = None;
         self.input_label_frame = None;
         self.csv_data = None;
+        self.show_data = None;
         self.current_data = None;
         self.input_label_frame = tk.LabelFrame(self.master, text="Input Data")
         self.input_label_frame.config(font=("Calibri", 14))
@@ -65,6 +66,8 @@ class Main(tk.Frame):
         self.inner_fields_frame_1 = tk.Frame(self.input_label_frame);
         self.inner_fields_frame_1.pack(side=tk.TOP, fill="x")
         self.__data_file_input = Input(self.inner_fields_frame_1, 'CSV File', action=self.select_file);
+
+
 
     def show_loading(self):
         self.loading_label = tk.Label(self.master, text='Loading... Please Wait..')
@@ -76,12 +79,24 @@ class Main(tk.Frame):
 
     def select_file(self):
         """Selects a file, get data, convert to dataframe and load it to table"""
+        self.clearBtns();
         self.__file_name = tk.filedialog.askopenfilename(initialdir=os.getcwd(), \
                                                 filetypes=(('CSV files', 'csv'),), title="Select Input CSV File.")
-        self.reset();
+        if self.__file_name is not None and self.__file_name is not '':
+            self.__data_file_input.set_data(self.__file_name)
+            self.load_btns();
+        else:
+            tk.messagebox.showerror(message="Please select a valid file.");
+            return;
+
+    def load_data(self):
+        if self.table_frame is not None:
+            print('destroyed');
+            self.table_frame.destroy();
         self.show_loading();
-        self.__data_file_input.set_data(self.__file_name)
         self.master.update();
+        self.csv_data = None;
+        self.current_data = None;
         try:
             self.csv_data = pd.read_csv(self.__file_name, low_memory=False)
             self.csv_data['date'] = pd.to_datetime((self.csv_data['date'] + " " + \
@@ -94,13 +109,16 @@ class Main(tk.Frame):
             self.hide_loading();
             self.csv_data = self.csv_data.drop(['time'], axis=1)
             self.current_data = self.csv_data;
-            self.load_btns()
+            self.filter(self.filter_value_1.get())
+
 
     def save_csv(self):
+        if self.current_data is None or len(self.current_data) == 0:
+            tk.messagebox.showerror(message="Load data before saving!");
+            return;
         filename = tk.filedialog.asksaveasfilename(initialdir=os.getcwd(), \
                     initialfile='Timeline.csv', filetypes=(('CSV files', 'csv'),), \
                                             title="Select Input CSV File.", defaultextension='*.csv')
-        print(filename);
         try:
             self.current_data.to_csv(filename, index=False, encoding='utf-8');
         except Exception as e:
@@ -111,18 +129,24 @@ class Main(tk.Frame):
 
     def reset(self):
         """Reset the data to the initial data loaded from the CSV"""
+        self.clearBtns();
+        self.show_loading();
+        self.master.update();
+        #if self.csv_data is not None:
+        #    self.current_data = self.csv_data;
+        self.current_data = None;
+        self.hide_loading();
+        if self.inner_fields_frame_2 is not None:
+            self.load_btns()
+
+    def clearBtns(self):
         if self.table_frame is not None:
             print('destroyed');
             self.table_frame.destroy();
         if self.inner_fields_frame_2 is not None:
             self.inner_fields_frame_2.destroy();
-        self.show_loading();
-        self.master.update();
-        if self.csv_data is not None:
-            self.current_data = self.csv_data;
-        self.hide_loading();
-        if self.inner_fields_frame_2 is not None:
-            self.load_btns()
+        if self.show_data is not None:
+            self.show_data.destroy();
 
     def load_btns(self):
         """load buttons like filter, search and query builder"""
@@ -131,23 +155,28 @@ class Main(tk.Frame):
         filter_label = tk.Label(self.inner_fields_frame_2, text='Filter Columns')
         filter_label.config(font=("Calibri", 12))
         filter_label.pack(side="left");
-        self.filter_value = tk.StringVar()
-        self.filter_value = tk.Entry(self.inner_fields_frame_2, textvariable=self.filter_value)
+        self.filter_value_1 = tk.StringVar()
+        self.filter_value = tk.Entry(self.inner_fields_frame_2, textvariable=self.filter_value_1)
         self.filter_value.config(font=("Calibri", 14), width=int(config_dict['entry_width']))
         self.filter_value.pack(side="left");
-        self.load_table();
-        filter = tk.Button(self.inner_fields_frame_2, text="Filter", \
-                           width=int(config_dict['button_width']), command=(lambda :self.filter(self.filter_value.get())))
-        filter.pack(side='left')
         help = tk.Button(self.inner_fields_frame_2, width=20, height=20, command=self.help_window)
         img = tk.PhotoImage(file="help.png", width=20, height=20)
         help.config(image=img)
         help.image = img;
         help.pack(side='left')
+
+        self.show_data = tk.Button(self.inner_fields_frame_2, text="Load data", \
+                                   width=int(config_dict['button_width']),
+                                   command=self.load_data)
+        self.show_data.pack(side='left', padx=10)
+        #self.load_table();
+        #filter = tk.Button(self.inner_fields_frame_2, text="Filter", \
+        #                   width=int(config_dict['button_width']), command=(lambda :self.filter(self.filter_value.get())))
+        #filter.pack(side='left')
         search = tk.Button(self.inner_fields_frame_2, text="Search", width = int(config_dict['button_width']), \
                            command=self.search_window)
         search.pack(side='left', padx=10)
-        reset = tk.Button(self.inner_fields_frame_2, text="Reset", width = int(config_dict['button_width']), \
+        reset = tk.Button(self.inner_fields_frame_2, text="Clear", width = int(config_dict['button_width']), \
                           command=self.reset)
         reset.pack(side='left', padx=10)
         save = tk.Button(self.inner_fields_frame_2, text="Save as CSV", width=int(config_dict['button_width']), \
@@ -224,27 +253,36 @@ class Main(tk.Frame):
 
     def search(self, srch_value):
         """Performs case insensitive search on data loaded from CSV"""
-        self.table_frame.destroy();
+        if self.current_data is None or len(self.current_data) == 0:
+            tk.messagebox.showerror(message="No data to search! Please load data first.");
+            return;
+        if srch_value is None or srch_value is '':
+            tk.messagebox.showerror(message="Enter a search value!");
+            return;
+        if self.table_frame:
+            self.table_frame.destroy();
         self.show_loading();
         self.master.update();
         self.search_win.destroy();
         mask = np.column_stack(\
-            [self.csv_data[col].astype(str).str.lower().astype(str).str.contains(srch_value.lower(), na=False) for col in self.csv_data])
-        self.current_data = self.csv_data.loc[mask.any(axis=1)]
+            [self.current_data[col].astype(str).str.lower().astype(str).str.contains(srch_value.lower(), na=False, regex=True) for col in self.current_data])
+        self.current_data = self.current_data.loc[mask.any(axis=1)]
         self.hide_loading();
         self.load_table();
 
     def filter(self, query):
         """Filter data based on the query"""
-        if query == '':
-            tk.messagebox.showerror(message="Empty query!");
-            return;
-        self.table_frame.destroy()
+        #if query == '':
+        #    tk.messagebox.showerror(message="Empty query!");
+        #    return;
+        if self.table_frame:
+            self.table_frame.destroy()
         self.show_loading();
         self.master.update();
         print('Query : ', query);
         try:
-            self.current_data = self.csv_data.query(query)
+            if query != '':
+                self.current_data = self.csv_data.query(query)
         except Exception as e:
             self.hide_loading();
             print('Exception ', e)
